requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
feedparser==6.0.11
zenrows==1.3.2
requests==2.31.0
html5lib
bs4==0.0.2
datetime==5.5
pandas==2.2.2
numpy==1.26.4
msrest==0.7.1
msrestazure
azure-core==1.30.1
azure-common==1.1.28
azure-identity==1.16.0
azure-storage-blob==12.19.1
azure-keyvault-secrets==4.8.0